calculation of temparature and accel fixed

+++ src/i2c_lsm9ds0/lsm9ds0.h
@@ -117,6 +117,6 @@ void lsm9ds0_init_sensor(uint32_t i2c, uint8_t sensor);
 void lsm9ds0_setup_temp(uint32_t i2c, uint8_t sensor);
 
 lsm9ds0Vector_t lsm9ds0_read_accel(uint32_t i2c, uint8_t sensor);
-uint16_t lsm9ds0_read_temp(uint32_t i2c, uint8_t sensor);
+float lsm9ds0_read_temp(uint32_t i2c, uint8_t sensor);
 
 #endif
